requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.1
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multidict==6.4.4
openai==1.82.0
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
